Fix error resulting from wrong dtype in tensor copies and numpy arrays types

+++ downscale_code/tests/test_helper.py
@@ -26,7 +26,7 @@ def close_equal(first, second):
 
 
 def test_img0():
-    return numpy.arange(8 * 8 * 3).reshape((8, 8, 3))
+    return numpy.arange(8 * 8 * 3).reshape((8, 8, 3)).astype(numpy.float64)
 
 
 def downscaled_img0():
